Add function to reset the OV580 module without resetting the MCU

+++ src/sensorcapture.cpp
@@ -604,13 +604,20 @@ bool SensorCapture::sendPing() {
     return true;
 }
 
-bool SensorCapture::resetSensorModule(int serial_number)
+bool SensorCapture::searchForConnectedDev(int* serial_number, unsigned short* found_pid)
 {
+    int in_serial_number;
+    if(serial_number==nullptr)
+         in_serial_number = 0;
+    else
+        in_serial_number = *serial_number;
+    int found_serial_number = 0;
+
     // ----> Search for connected device
     struct hid_device_info *devs, *cur_dev;
 
     if (hid_init()==-1)
-        return 0;
+        return false;
 
     devs = hid_enumerate(SL_USB_VENDOR, 0x0);
     cur_dev = devs;
@@ -626,11 +633,12 @@ bool SensorCapture::resetSensorModule(int serial_number)
         sn_str = wstr2str( cur_dev->serial_number );
         int sn = std::stoi( sn_str );
 
-        if(serial_number==0 || sn==serial_number)
+        if(in_serial_number==0 || sn==in_serial_number)
         {
             if( pid==SL_USB_PROD_MCU_ZED2_REVA || pid==SL_USB_PROD_MCU_ZED2i_REVA)
             {
                 found = true;
+                found_serial_number = sn;
                 break;
             }
             else
@@ -638,7 +646,7 @@ bool SensorCapture::resetSensorModule(int serial_number)
                 std::string msg = "The reset function works only for ZED2/ZED2i camera models.";
                 std::cerr << msg << std::endl;
 
-                if(serial_number==0)
+                if(in_serial_number==0)
                     continue;
                 else
                     return false;
@@ -651,17 +659,33 @@ bool SensorCapture::resetSensorModule(int serial_number)
     hid_free_enumeration(devs);
     // <---- Search for connected device
 
-    if(!found)
+    if(!found) {
+        return false;
+    }
+
+    if(serial_number)
+        *serial_number = found_serial_number;
+    if(found_pid)
+        *found_pid = pid;
+    return true;
+}
+
+bool SensorCapture::resetSensorModule(int serial_number)
+{
+    int found_sn = serial_number;
+    unsigned short pid;
+    bool res = searchForConnectedDev(&found_sn, &pid);
+    if(!res)
     {
         std::string msg;
         if(serial_number!=0)
         {
-            msg = "Unable to find the Sensor Module with serial number ";
+            msg = "[sl_oc::sensors::SensorCapture] WARNING: Sensors Module reset failed. Unable to find the Sensor Module with serial number ";
             msg += std::to_string(serial_number);
         }
         else
         {
-            msg = "Unable to find the Sensor Module of a ZED2 camera. Please verify the USB connection.";
+            msg = "[sl_oc::sensors::SensorCapture] WARNING: Sensors Module reset failed. Unable to find the Sensor Module of a ZED2 camera. Please verify the USB connection.";
         }
 
         std::cerr << msg << std::endl;
@@ -669,6 +693,7 @@ bool SensorCapture::resetSensorModule(int serial_number)
         return false;
     }
 
+    std::string sn_str = std::to_string(found_sn);
     std::wstring wide_sn_string = std::wstring(sn_str.begin(), sn_str.end());
     const wchar_t* wsn = wide_sn_string.c_str();
 
@@ -687,14 +712,72 @@ bool SensorCapture::resetSensorModule(int serial_number)
     buf[1] = static_cast<unsigned char>(usb::RQ_CMD_RST);
 
     hid_send_feature_report(devHandle, buf, 2);
-    sleep(2); // Wait for MCU to reboot
+    // Note: cannot verify the return value of the `hid_send_feature_report` command because the MCU is suddenly reset
+    // and it cannot return a valid value
+
+    sleep(2); // Wait for MCU and OV580 to reboot
+
+    std::cerr << "[sl_oc::sensors::SensorCapture] INFO: Sensors Module reset successful" << std::endl;
 
     return true;
 }
 
 bool SensorCapture::resetVideoModule(int serial_number)
 {
+    int found_sn = serial_number;
+    unsigned short pid;
+    bool res = searchForConnectedDev(&found_sn, &pid);
+    if(!res)
+    {
+        std::string msg;
+        if(serial_number!=0)
+        {
+            msg = "[sl_oc::sensors::SensorCapture] WARNING: Video Module reset failed. Unable to find the Sensor Module with serial number ";
+            msg += std::to_string(serial_number);
+        }
+        else
+        {
+            msg = "[sl_oc::sensors::SensorCapture] WARNING: Video Module reset failed. Unable to find the Sensor Module of a ZED2 camera. Please verify the USB connection.";
+        }
 
+        std::cerr << msg << std::endl;
+
+        return false;
+    }
+
+    std::string sn_str = std::to_string(found_sn);
+    std::wstring wide_sn_string = std::wstring(sn_str.begin(), sn_str.end());
+    const wchar_t* wsn = wide_sn_string.c_str();
+
+    hid_device* devHandle = hid_open(SL_USB_VENDOR, pid, wsn );
+
+    if(!devHandle)
+    {
+        std::string msg = "Unable to open the MCU HID device";
+        std::cerr << msg << std::endl;
+
+        return false;
+    }
+
+    usb::OV580CmdStruct cmd;
+    cmd.struct_id = usb::REP_ID_OV580_CMD;
+    cmd.cmd = usb::OV580_CMD_RESET;
+    cmd.info=0;
+
+    unsigned char buf[65];
+    memcpy(buf, &(cmd.struct_id), sizeof(usb::OV580CmdStruct));
+
+    int ret = hid_send_feature_report(devHandle, buf, sizeof(usb::OV580CmdStruct));
+    hid_close(devHandle);
+
+    if(ret!=sizeof(usb::OV580CmdStruct)) {
+        std::cerr << "[sl_oc::sensors::SensorCapture] INFO: Video Module reset failed" << std::endl;
+        return false;
+    }
+
+    sleep(2); // Wait for OV580 to reboot
+
+    std::cerr << "[sl_oc::sensors::SensorCapture] INFO: Video Module reset successful" << std::endl;
     return true;
 }
